Removed links table from process
Links table was the biggest table (35GB) which required over 1h to
construct. It was only used to construct its compact version --
link edges table. Now the latter is constructed directly, so there
the links table is removed, saving space and time.

Also some other minor changes in this commit because I was too lazy
to split.